并行网关&&节点变量处理

+++ workflow/model/bytearry.go
@@ -2,7 +2,7 @@ package model
 
 //流程源数据
 type Bytearry struct {
-	Id           int    `xorm:"id" xorm:"pk autoincr"`
+	Id           int64  `xorm:"id" xorm:"pk autoincr"`
 	Key          string `xorm:"key"`
 	Name         string `xorm:"name"`
 	Version      int    `xorm:"version"`
